+++ src/components/checkPassword/styles.ts
